The tasks stored in Hive boxes are not easily drawn and it takes time. So further program cannot be proceded until data is retrived. Here we cannot use async function hence we use futureBuilder function
Even the initstate function will also not work.

Storing few tasks using add function.

The values in the database are stored in terms of box which is converted into map using a function

The ListView.Builder funtion provides the platform to display tasks as they are present in database taking the arguments as shown.

The checkboxes and text decoration are operated based on the boolean value of task completion